refactored and combined same-route views

+++ app/api_1_0/bucketlist_items.py
@@ -39,11 +39,10 @@ def create_bucketlist_item(id):
     }), 201
 
 
-
-@api.route('/bucketlists/<int:id>/items/<int:item_id>', methods = ['PUT'])
+@api.route('/bucketlists/<int:id>/items/<int:item_id>', methods = ['PUT', 'DELETE'])
 @jwt_required()
-def update_bucketlist_item(id, item_id):
-    """ updates an existing bucketlist item. 
+def manage_bucketlist_item(id, item_id):
+    """ updates or deletes an existing bucketlist item. 
     """
     # get the bucketlist:
     try:
@@ -56,51 +55,37 @@ def update_bucketlist_item(id, item_id):
         bucketlist_item = BucketlistItem.get_bucketlist_item(bucketlist, item_id)
     except Exception, e:
         return not_found(e.message)
+    
+    if request.method == 'PUT':
+        # update it with the json values:
+        json_bucketlist_item = request.json
+        name = json_bucketlist_item.get('name')
+        done = json_bucketlist_item.get('done')
         
-    # update it with the json values:
-    json_bucketlist_item = request.json
-    name = json_bucketlist_item.get('name')
-    done = json_bucketlist_item.get('done')
-    if name:
-        bucketlist_item.name = name
-    if isinstance(done, bool):
-        bucketlist_item.done = done
+        if name:
+            bucketlist_item.name = name
+        if isinstance(done, bool):
+            bucketlist_item.done = done
 
-    # save the bucketlist_item to the db:
-    db.session.add(bucketlist_item)
-    db.session.commit()
+        # save the bucketlist_item to the db:
+        db.session.add(bucketlist_item)
+        db.session.commit()
 
-    # return the json response:
-    return jsonify({
-        "bucketlist_item": bucketlist_item.to_json(),
-        "bucketlist_url": url_for('api.get_bucketlist', id=bucketlist.id, _external=True)
-    }), 200
+        # return the json response:
+        return jsonify({
+            "bucketlist_item": bucketlist_item.to_json(),
+            "bucketlist_url": url_for('api.get_bucketlist', id=bucketlist.id, _external=True)
+        }), 200
 
+    elif request.method == 'DELETE':
+         # delete the bucketlist from the db:
+        db.session.delete(bucketlist_item)
+        db.session.commit()
 
+        # return the json response:
+        return jsonify({
+            "status": "deleted",
+            "bucketlist_url": url_for('api.get_bucketlist', id=bucketlist.id, _external=True)
+        }), 200
 
-@api.route('/bucketlists/<int:id>/items/<int:item_id>', methods = ['DELETE'])
-@jwt_required()
-def delete_bucketlist_item(id, item_id):
-    """ deletes an existing bucketlist-item. 
-    """
-    # get the bucketlist:
-    try:
-        bucketlist = Bucketlist.get_user_bucketlist(current_identity, id)
-    except Exception, e:
-        return not_found(e.message)
-
-    # get bucketlist-item:
-    try:
-        bucketlist_item = BucketlistItem.get_bucketlist_item(bucketlist, item_id)
-    except Exception, e:
-        return not_found(e.message)
-
-    # delete the bucketlist from the db:
-    db.session.delete(bucketlist_item)
-    db.session.commit()
-
-    # return the json response:
-    return jsonify({
-        "status": "deleted",
-        "bucketlist_url": url_for('api.get_bucketlist', id=bucketlist.id, _external=True)
-    }), 200 
+
